app/services/db.py
import os
from dotenv import load_dotenv
from sqlmodel import SQLModel, create_engine, Session

load_dotenv()

Database_URL = os.getenv("DATABASE_URL")

print(Database_URL)

engine = create_engine(Database_URL, echo=True)

def init_db():
    SQLModel.metadata.create_all(engine)
    
def get_session():
    with Session(engine) as session:
        yield session