requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
bcrypt
pyjwt
uuid6
sqlmodel
python-dotenv